refactor logic to authentication service and used es6 syntax
